requirements.txt
discord
python-dotenv
discord.py[voice]